Refactoring PropertyMethodArbitraryResolver
